Moved `.encode()` and `.decode_raw()` from models to `utils` module
